refactor: SectionRequest 패키지 변경
